+++ handler.go
@@ -8,11 +8,6 @@ import (
 // HandlerFunc is the accepted function to use in a Handler.
 type HandlerFunc func(w http.ResponseWriter, r *http.Request) (Responder, error)
 
-// ErrorLoggerFunc is a function that logs an error in a request.
-// Since its goal is only debugging errors, it runs in a different goroutine
-// and passes a deep copy of the request as the first argument.
-type ErrorLoggerFunc func(r http.Request, err error)
-
 // ErrorHandlerFunc is a function used to handle runtime errors.
 type ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
 
@@ -22,16 +17,12 @@ type MarshallerFunc func(v interface{}) ([]byte, error)
 
 // RuntimeErrorFunc is a function used to return an
 // Error interface which will be sent as response.
-type RuntimeErrorFunc func(r http.Request) Error
+type RuntimeErrorFunc func(r *http.Request) Error
 
 var (
 	// DefaultContentType is the default Content-Type MIME
 	// type a Handler uses when it is created.
 	DefaultContentType = "application/json"
-	// DefaultErrorLoggerFunc is the default function used
-	// by a Handler when it is created to log information
-	// when an error occurs.
-	DefaultErrorLoggerFunc ErrorLoggerFunc
 	// DefaultErrorHandlerFunc is the default function for
 	// handling runtime errors.
 	DefaultErrorHandlerFunc ErrorHandlerFunc
@@ -54,7 +45,6 @@ var (
 type Handler struct {
 	ContentType      string
 	HandlerFunc      HandlerFunc
-	ErrorLoggerFunc  ErrorLoggerFunc
 	ErrorHandlerFunc ErrorHandlerFunc
 	MarshallerFunc   MarshallerFunc
 	ErrMsg           string
@@ -67,7 +57,6 @@ func New(hfunc HandlerFunc) *Handler {
 	return &Handler{
 		ContentType:      DefaultContentType,
 		HandlerFunc:      hfunc,
-		ErrorLoggerFunc:  DefaultErrorLoggerFunc,
 		ErrorHandlerFunc: DefaultErrorHandlerFunc,
 		MarshallerFunc:   DefaultMarshallerFunc,
 		ErrCode:          DefaultErrCode,
@@ -93,8 +82,6 @@ func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", h.ContentType)
 
 	if err != nil {
-		h.logError(r, err)
-
 		switch e := err.(type) {
 		case Error:
 			if err = h.write(w, e.Body(), e.Status()); err != nil {
@@ -114,28 +101,19 @@ func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	err = h.write(w, res.Body(), res.Status())
 
 	if err != nil {
-		h.logError(r, err)
 		h.handleError(w, r, err)
 	}
 }
 
 func (h *Handler) handleError(w http.ResponseWriter, r *http.Request, err error) {
 	if h.RuntimeErrorFunc != nil {
-		runtimeErr := h.RuntimeErrorFunc(*r)
+		runtimeErr := h.RuntimeErrorFunc(r)
 
 		if err = h.write(w, runtimeErr, runtimeErr.Status()); err != nil {
-			h.logError(r, err)
-
 			if h.ErrorHandlerFunc != nil {
 				h.ErrorHandlerFunc(w, r, err)
 			}
 		}
-	}
-}
-
-func (h *Handler) logError(r *http.Request, err error) {
-	if h.ErrorLoggerFunc != nil {
-		h.ErrorLoggerFunc(*r, err)
 	}
 }