+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Runtime error handler.
 
+### Removed
+- Logger function support.
+
 ### Changed
 - `Error` now implements both `error` and `Responder`.